Island: Move Logout to authentication_service
