+++ monkey/monkey_island/cc/resources/auth/__init__.py
@@ -1,2 +1,1 @@
-from .logout import Logout
 from .registration_status import RegistrationStatus
