+++ monkey/monkey_island/cc/services/authentication_service/flask_resources/logout.py
